feat/#7/요청받는 id값이 유효한지 확인하는 로직 추가

+++ src/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { Users } from '@prisma/client';
 import { PrismaService } from '../prisma/prisma.service';
 import { CreateUserDto } from './dto/create-user.dto';
@@ -13,7 +13,11 @@ export class UsersService {
   }
 
   async getUser(id: number): Promise<Users> {
-    return await this.prisma.users.findUnique({ where: { id } });
+    const user = await this.prisma.users.findUnique({ where: { id } });
+    if (!user) {
+      throw new NotFoundException(`ID ${id} not found`);
+    }
+    return user;
   }
 
   async createUser(createUserData: CreateUserDto): Promise<Users> {
@@ -21,13 +25,23 @@ export class UsersService {
   }
 
   async updateUser(id: number, updateUserData: UpdateUserDto): Promise<Users> {
-    return await this.prisma.users.update({
+    if (!(await this.prisma.users.findUnique({ where: { id } }))) {
+      throw new NotFoundException(`ID ${id} not found`);
+    }
+
+    const user = await this.prisma.users.update({
       where: { id },
       data: updateUserData,
     });
+
+    return user;
   }
 
   async deleteUser(id: number): Promise<Users> {
+    if (!(await this.prisma.users.findUnique({ where: { id } }))) {
+      throw new NotFoundException(`ID ${id} not found`);
+    }
+
     return this.prisma.users.delete({ where: { id } });
   }
 }
